+++ interfaces/web/models/dash.py
@@ -3,19 +3,13 @@ import time
 import plotly
 import plotly.graph_objs as go
 
-from config.cst import TimeFrames, EvaluatorMatrixTypes, PriceIndexes, CONFIG_EVALUATOR, CONFIG_TENTACLES_KEY, \
-    TENTACLE_PACKAGE_DESCRIPTION
+from config.cst import TimeFrames, EvaluatorMatrixTypes, PriceIndexes
 from evaluator.evaluator_matrix import EvaluatorMatrix
 from interfaces import get_reference_market, get_bot, set_default_time_frame, get_default_time_frame
 from interfaces.trading_util import get_portfolio_current_value, get_trades_by_times_and_prices
 from interfaces.web import add_to_matrix_history, get_matrix_history, add_to_symbol_data_history, \
     add_to_portfolio_value_history, get_portfolio_value_history, TIME_AXIS_TITLE, get_symbol_data_history
 from tools.symbol_util import split_symbol
-from tools.config_manager import ConfigManager
-from tools.tentacle_manager.tentacle_package_manager import TentaclePackageManager
-from tools.tentacle_manager.tentacle_package_util import get_is_url, get_package_name, \
-    get_octobot_tentacle_public_repo, get_package_description_with_adaptation
-from tools.tentacle_manager.tentacle_manager import TentacleManager
 from trading.trader.portfolio import Portfolio
 
 
@@ -220,57 +214,3 @@ def get_evaluator_graph_in_matrix_history(symbol,
                                title="Buy or sell")
                 )}
     return None
-
-
-def get_evaluator_config():
-    return get_bot().get_config()[CONFIG_EVALUATOR]
-
-
-def get_evaluator_startup_config():
-    return get_bot().get_startup_config()[CONFIG_EVALUATOR]
-
-
-def reset_evaluator_config():
-    return update_evaluator_config(get_evaluator_startup_config())
-
-
-def update_evaluator_config(new_config):
-    current_config = get_bot().get_config()[CONFIG_EVALUATOR]
-    try:
-        ConfigManager.update_evaluator_config(new_config, current_config)
-        return True
-    except Exception:
-        return False
-
-
-def get_tentacles_packages():
-    default_tentacles_repo_desc_file = get_octobot_tentacle_public_repo()
-    default_tentacles_repo = get_octobot_tentacle_public_repo(False)
-    packages = {
-        default_tentacles_repo: get_package_name(default_tentacles_repo_desc_file,
-                                                 get_is_url(default_tentacles_repo_desc_file))
-    }
-    for tentacle_package in get_bot().get_config()[CONFIG_TENTACLES_KEY]:
-        packages[tentacle_package] = get_package_name(tentacle_package, get_is_url(tentacle_package))
-    return packages
-
-
-def get_tentacles_package_description(path_or_url):
-    try:
-        return get_package_description_with_adaptation(path_or_url)[TENTACLE_PACKAGE_DESCRIPTION]
-    except Exception:
-        return None
-
-
-def register_and_install(path_or_url):
-    config = get_bot().get_config()
-    config[CONFIG_TENTACLES_KEY].append(path_or_url)
-    #TODO update config.json
-    
-    tentacles_manager = TentacleManager(config)
-    tentacles_manager.install_tentacle_package(path_or_url, True)
-    return True
-
-
-def get_tentacles():
-    return TentaclePackageManager.get_installed_modules()
